requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.9.0
pydantic-settings==2.5.0
python-multipart==0.0.9
qdrant-client==1.11.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.35
openai==1.45.0
httpx==0.27.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
sentence-transformers==3.0.1
numpy==1.26.4
tenacity==9.0.0
sentry-sdk==2.15.0
prometheus-client==0.21.0
